[feat] #9 이미지 프로세서 로직 및 메탈 이미지 뷰 구성

+++ Filtee/Feature/FilterDetail/FilterDetailView.swift
@@ -376,7 +376,7 @@ private extension FilterDetailView {
     
     func filterValue(
         _ resource: ImageResource,
-        value: Double
+        value: Float
     ) -> some View {
         VStack(spacing: 4) {
             Image(resource)
